refactor(api): move jwt.ts to auth/jwt.ts
